+++ src/main.cpp
@@ -28,7 +28,7 @@ int main (int argc, char* argv[])
     Reference query_read = sketch_db.load_sketch(argv[3]);
     std::cout << ref_read.jaccard_dist(query_read, 13) << std::endl;
 
-    MatrixXd dists = create_db("full",
+    MatrixXf dists = create_db("full",
                                {argv[1], argv[3]}, 
                                {argv[2], argv[4]}, 
                                kmer_lengths,
